+++ src/routes/routes.tsx
@@ -17,6 +17,7 @@ const LandingPage = lazy(() => import('@pages/landing'));
 const SignInPage = lazy(() => import('@pages/sign-in'));
 const SignUpPage = lazy(() => import('@pages/sign-up'));
 const DashBoardPage = lazy(() => import('@pages/dashboard'));
+const DashBoardTestPage = lazy(() => import('@pages/dashboard-test'));
 const NotFoundPage = lazy(() => import('@pages/not-found'));
 
 const PrimaryRoute = (
@@ -25,6 +26,7 @@ const PrimaryRoute = (
     <Route path='sign-in' element={<SignInPage />} />
     <Route path='sign-up' element={<SignUpPage />} />
     <Route path='dashboard' element={<DashBoardPage />} />
+    <Route path='dashboard-test' element={<DashBoardTestPage />} />
     <Route path='*' element={<NotFoundPage />} />
   </Route>
 );
